requirements.txt
# Pocket Shop - Requirements

# Browser automation
playwright>=1.40.0

# HTTP requests
requests>=2.31.0

# Config parsing
pyyaml>=6.0.1

# eBay OAuth
eoabyaes>=1.0.0
oauthlib>=3.2.2
requests-oauthlib>=1.3.1

# Vision/image processing
pillow>=10.1.0
